Initial submit of new library providing Juniper SRX filter output support. The Juniper SRX support is BETA, and still under development.
Special thanks to Rob Ankeny (robankeny@google.com) for his time and efforts in
development of this generator library.

Note: As juniper SRX is a fully stateful platform, options such as
'tcp-established' are not readily meaningful/usable at this time.  As testing
and development continues we will work to implement more keyword and option
support as time and resources allow.